refactor: useIdolChunks 훅 생성
